Move locks retrieval to the GET action
