+++ core/src/test/java/google/registry/ui/server/registrar/RegistryLockGetActionTest.java
@@ -15,28 +15,40 @@
 package google.registry.ui.server.registrar;
 
 import static com.google.common.truth.Truth.assertThat;
+import static google.registry.request.auth.AuthenticatedRegistrarAccessor.Role.OWNER;
+import static google.registry.testing.AppEngineRule.makeRegistrar2;
+import static google.registry.testing.AppEngineRule.makeRegistrarContact3;
+import static google.registry.testing.DatastoreHelper.persistResource;
 import static google.registry.testing.JUnitBackports.assertThrows;
-import static org.mockito.Mockito.when;
+import static javax.servlet.http.HttpServletResponse.SC_FORBIDDEN;
+import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
 
 import com.google.api.client.http.HttpStatusCodes;
 import com.google.appengine.api.users.User;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSetMultimap;
 import com.google.gson.Gson;
+import google.registry.model.registry.RegistryLockDao;
+import google.registry.model.transaction.JpaTransactionManagerRule;
 import google.registry.request.Action.Method;
 import google.registry.request.auth.AuthLevel;
 import google.registry.request.auth.AuthResult;
-import google.registry.request.auth.AuthenticatedRegistrarAccessor.RegistrarAccessDeniedException;
+import google.registry.request.auth.AuthenticatedRegistrarAccessor;
 import google.registry.request.auth.UserAuthInfo;
+import google.registry.schema.domain.RegistryLock;
+import google.registry.schema.domain.RegistryLock.Action;
 import google.registry.testing.AppEngineRule;
 import google.registry.testing.FakeResponse;
+import java.util.Map;
 import java.util.Optional;
+import java.util.UUID;
+import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
@@ -47,46 +59,93 @@ public final class RegistryLockGetActionTest {
   private static final Gson GSON = new Gson();
 
   @Rule public final AppEngineRule appEngineRule = AppEngineRule.builder().withDatastore().build();
+
+  @Rule
+  public final JpaTransactionManagerRule jpaTmRule =
+      new JpaTransactionManagerRule.Builder().build();
+
   @Rule public final MockitoRule mocks = MockitoJUnit.rule();
 
   private final FakeResponse response = new FakeResponse();
-  private final User user = new User("marla.singer@example.com", "gmail.com", "12345");
+  private final User user = new User("Marla.Singer@crr.com", "gmail.com", "12345");
 
   private AuthResult authResult;
+  private AuthenticatedRegistrarAccessor accessor;
   private RegistryLockGetAction action;
-
-  @Mock ExistingRegistryLocksRetriever retriever;
 
   @Before
   public void setup() {
+    jpaTmRule.getTxnClock().setTo(DateTime.parse("2000-06-08T22:00:00.0Z"));
     authResult = AuthResult.create(AuthLevel.USER, UserAuthInfo.create(user, false));
+    accessor =
+        AuthenticatedRegistrarAccessor.createForTesting(
+            ImmutableSetMultimap.of(
+                "TheRegistrar", OWNER,
+                "NewRegistrar", OWNER));
     action =
         new RegistryLockGetAction(
-            Method.GET, response, authResult, retriever, Optional.of("clientId"));
+            Method.GET, response, accessor, authResult, Optional.of("TheRegistrar"));
   }
 
   @Test
-  public void testSuccess_retrievesLocks() throws Exception {
-    ImmutableMap<String, Object> resultMap =
-        ImmutableMap.of(
-            "lockEnabledForContact",
-            true,
-            "email",
-            "marla.singer@example.com",
-            "locks",
-            ImmutableList.of(),
-            "clientId",
-            "clientId");
-    when(retriever.getLockedDomainsMap("clientId")).thenReturn(resultMap);
+  public void testSuccess_retrievesLocks() {
+    RegistryLock regularLock =
+        new RegistryLock.Builder()
+            .setRepoId("repoId")
+            .setDomainName("example.test")
+            .setRegistrarId("TheRegistrar")
+            .setAction(Action.LOCK)
+            .setVerificationCode(UUID.randomUUID().toString())
+            .setRegistrarPocId("johndoe@theregistrar.com")
+            .setCompletionTimestamp(jpaTmRule.getTxnClock().nowUtc())
+            .build();
+    jpaTmRule.getTxnClock().advanceOneMilli();
+    RegistryLock adminLock =
+        new RegistryLock.Builder()
+            .setRepoId("repoId")
+            .setDomainName("adminexample.test")
+            .setRegistrarId("TheRegistrar")
+            .setAction(Action.LOCK)
+            .setVerificationCode(UUID.randomUUID().toString())
+            .isSuperuser(true)
+            .setCompletionTimestamp(jpaTmRule.getTxnClock().nowUtc())
+            .build();
+    RegistryLock incompleteLock =
+        new RegistryLock.Builder()
+            .setRepoId("repoId")
+            .setDomainName("incomplete.test")
+            .setRegistrarId("TheRegistrar")
+            .setAction(Action.LOCK)
+            .setVerificationCode(UUID.randomUUID().toString())
+            .setRegistrarPocId("johndoe@theregistrar.com")
+            .build();
+
+    RegistryLockDao.save(regularLock);
+    RegistryLockDao.save(adminLock);
+    RegistryLockDao.save(incompleteLock);
+
     action.run();
     assertThat(response.getStatus()).isEqualTo(HttpStatusCodes.STATUS_CODE_OK);
-    assertThat(response.getPayload())
-        .isEqualTo(
-            GSON.toJson(
-                ImmutableMap.of(
-                    "status", "SUCCESS",
-                    "message", "Successful locks retrieval",
-                    "results", ImmutableList.of(resultMap))));
+    assertThat(GSON.fromJson(response.getPayload(), Map.class))
+        .containsExactly(
+            "status", "SUCCESS",
+            "message", "Successful locks retrieval",
+            "results",
+                ImmutableList.of(
+                    ImmutableMap.of(
+                        "lockEnabledForContact", true,
+                        "email", "Marla.Singer@crr.com",
+                        "clientId", "TheRegistrar",
+                        "locks",
+                        ImmutableList.of(
+                            ImmutableMap.of(
+                                "fullyQualifiedDomainName", "example.test",
+                                "lockedTime", "2000-06-08T22:00:00.000Z",
+                                "lockedBy", "johndoe@theregistrar.com"),
+                            ImmutableMap.of(
+                                "fullyQualifiedDomainName", "adminexample.test",
+                                "lockedTime", "2000-06-08T22:00:00.001Z",
+                                "lockedBy", "admin")))));
   }
 
   @Test
@@ -114,13 +173,84 @@ public final class RegistryLockGetActionTest {
   }
 
   @Test
-  public void testFailure_noRegistrarAccess() throws Exception {
-    String errorMessage =
-        String.format(
-            "%s doesn't have access to registrar clientId", authResult.userIdForLogging());
-    when(retriever.getLockedDomainsMap("clientId"))
-        .thenThrow(new RegistrarAccessDeniedException(errorMessage));
+  public void testFailure_noRegistrarAccess() {
+    accessor = AuthenticatedRegistrarAccessor.createForTesting(ImmutableSetMultimap.of());
+    action =
+        new RegistryLockGetAction(
+            Method.GET, response, accessor, authResult, Optional.of("TheRegistrar"));
     action.run();
-    assertThat(response.getStatus()).isEqualTo(HttpStatusCodes.STATUS_CODE_FORBIDDEN);
+    assertThat(response.getStatus()).isEqualTo(SC_FORBIDDEN);
+  }
+
+  @Test
+  public void testSuccess_readOnlyAccessForOtherUsers() {
+    // If lock is not enabled for a user, this should be read-only
+    persistResource(
+        makeRegistrarContact3().asBuilder().setAllowedToSetRegistryLockPassword(true).build());
+    action.run();
+    assertThat(GSON.fromJson(response.getPayload(), Map.class).get("results"))
+        .isEqualTo(
+            ImmutableList.of(
+                ImmutableMap.of(
+                    "lockEnabledForContact",
+                    false,
+                    "email",
+                    "Marla.Singer@crr.com",
+                    "clientId",
+                    "TheRegistrar",
+                    "locks",
+                    ImmutableList.of())));
+  }
+
+  @Test
+  public void testSuccess_lockAllowedForAdmin() throws Exception {
+    // Locks are allowed for admins even when they're not enabled for the registrar
+    persistResource(makeRegistrar2().asBuilder().setRegistryLockAllowed(false).build());
+    authResult = AuthResult.create(AuthLevel.USER, UserAuthInfo.create(user, true));
+    action =
+        new RegistryLockGetAction(
+            Method.GET, response, accessor, authResult, Optional.of("TheRegistrar"));
+    action.run();
+    assertThat(GSON.fromJson(response.getPayload(), Map.class).get("results"))
+        .isEqualTo(
+            ImmutableList.of(
+                ImmutableMap.of(
+                    "lockEnabledForContact",
+                    true,
+                    "email",
+                    "Marla.Singer@crr.com",
+                    "clientId",
+                    "TheRegistrar",
+                    "locks",
+                    ImmutableList.of())));
+  }
+
+  @Test
+  public void testFailure_lockNotAllowedForRegistrar() {
+    // The UI shouldn't be making requests where lock isn't enabled for this registrar
+    action =
+        new RegistryLockGetAction(
+            Method.GET, response, accessor, authResult, Optional.of("NewRegistrar"));
+    action.run();
+    assertThat(response.getStatus()).isEqualTo(SC_INTERNAL_SERVER_ERROR);
+  }
+
+  @Test
+  public void testFailure_accessDenied() {
+    accessor = AuthenticatedRegistrarAccessor.createForTesting(ImmutableSetMultimap.of());
+    action =
+        new RegistryLockGetAction(
+            Method.GET, response, accessor, authResult, Optional.of("TheRegistrar"));
+    action.run();
+    assertThat(response.getStatus()).isEqualTo(SC_FORBIDDEN);
+  }
+
+  @Test
+  public void testFailure_badRegistrar() {
+    action =
+        new RegistryLockGetAction(
+            Method.GET, response, accessor, authResult, Optional.of("SomeBadRegistrar"));
+    action.run();
+    assertThat(response.getStatus()).isEqualTo(SC_FORBIDDEN);
   }
 }
